backend/requirements.txt
asgiref==3.4.1
Django==2.2.16
django-filter==2.4.0
django-extra-fields==3.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.19.2
djoser==2.1.0
django-debug-toolbar==3.2.4
Pillow==9.1.0
